fastAPI_scraper/requirements.txt
fastapi
uvicorn
beautifulsoup4
requests
lxml
playwright
python-dotenv
